pyproject.toml
[project]
name = "context-engineering-xslt"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "behave>=1.3.0",
    "fastmcp>=2.0.0",
    "lxml>=6.0.0",
    "pydantic>=2.0.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.0.0",
    "ruff>=0.13.2",
    "saxonche>=12.8.0",
]
